+++ apps/formio-server/src/hooks/settings.js
@@ -103,6 +103,7 @@ module.exports = function(app) {
       configFormio: require('./alter/configFormio'),
       loadRevision: require('./alter/loadRevision')(app),
       parentProjectSettings: require('./alter/parentProjectSettings')(app),
+      rawDataAccess: require('./alter/rawDataAccess'),
       actionSettingsForm: require('./alter/actionSettingsForm')(app),
       saveSubmission: require('./alter/saveSubmission')(app),
       schemaIndex(index) {