+++ apps/formio-server/Changelog.md
@@ -13,6 +13,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - FIO-5728: checking for unique email when adding user to team
  - FIO-7236: move to deprecate random keygen
  - FIO-7397: fixed server validation errors for reporting ui config form
+ - FIO-7167: Isolated vm
+ - FIO-7423: Fortis Payment Gateway Feature
+ - FIO-7400: fixed an issue where lookup operator with pipeline gives an уrror in DocumentDB
+ - hotfix for failing tests
+ - FIO 7239: add support for AWS S3 multipart uploads
+ - FIO-7371: Adds hook that would allow to retrieve raw DB data
+ - FIO-7391: refactored reporting-ui configuration form
+ - FIO-7423: Failed to search current payment info in server
  
 ## 8.4.0-rc.7
 ### Changed